Develop ver.0.3.11

+++ Project/EngineLayer/PostEffectManagement/RandomEffect.cpp
@@ -68,8 +68,8 @@ void RandomEffect::Apply(ID3D12GraphicsCommandList* commandList, uint32_t srvInd
 	const uint32_t threadGroupSizeY = 8;
 
 	// レンダーターゲットの解像度（仮に 1280x720）
-	uint32_t width = randomSetting_->textureSize.x; // ウィンドウの幅
-	uint32_t height = randomSetting_->textureSize.y; // ウィンドウの高さ
+	uint32_t width = static_cast<uint32_t>(randomSetting_->textureSize.x); // ウィンドウの幅
+	uint32_t height = static_cast<uint32_t>(randomSetting_->textureSize.y); // ウィンドウの高さ
 
 	uint32_t groupCountX = (width + threadGroupSizeX - 1) / threadGroupSizeX;
 	uint32_t groupCountY = (height + threadGroupSizeY - 1) / threadGroupSizeY;
